[OBP-242]: Moves files to support new webpack config.
