+++ ingest/bin/index-migrator.rb
