--- ingest/lambdas/bulk-ingester/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "bulk-indexer",
-  "version": "1.0.0",
-  "description": "",
-  "main": "handler.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "p-map": "^4.0.0"
-  }
-}
