+++ ingest/lambdas/index-rectifier.js
