--- ingest/lambdas/index-rectifier/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "index-rectifier",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index-rectifier.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@aws-crypto/sha256-browser": "^2.0.0",
-    "@aws-sdk/client-sns": "^3.39.0",
-    "@aws-sdk/credential-provider-node": "^3.38.0",
-    "@aws-sdk/node-http-handler": "^3.38.0",
-    "@aws-sdk/protocol-http": "^3.38.0",
-    "@aws-sdk/signature-v4": "^3.38.0",
-    "got": "^11.8.2",
-    "p-map": "^4.0.0"
-  },
-  "devDependencies": {
-    "aws-sdk-client-mock": "^0.5.5",
-    "nock": "^13.1.4"
-  }
-}
