+++ lib/dspace-client.js
