+++ lib/dspace-client.test.js
@@ -1,6 +1,6 @@
 const got = require('got');
 
-const dspaceClient = require('../dspace-client');
+const dspaceClient = require('./dspace-client');
 
 jest.mock('got');
