+++ lib/open-search-client.js
