+++ lib/open-search-client.test.js
@@ -1,5 +1,5 @@
 const nock = require('nock');
-const osClient = require('../open-search-client');
+const osClient = require('./open-search-client');
 
 describe('open-search-client', () => {
   describe('openScroll', () => {
