+++ package.json
@@ -46,9 +46,17 @@
     "@aws-cdk/aws-ssm": "^1.129.0",
     "@aws-cdk/core": "^1.129.0",
     "@aws-cdk/pipelines": "^1.129.0",
+    "@aws-crypto/sha256-browser": "^2.0.0",
     "@aws-sdk/client-cloudformation": "^3.37.0",
     "@aws-sdk/client-s3": "^3.37.0",
+    "@aws-sdk/client-sns": "^3.40.0",
     "@aws-sdk/client-ssm": "^3.37.0",
+    "@aws-sdk/credential-provider-node": "^3.40.0",
+    "@aws-sdk/node-http-handler": "^3.40.0",
+    "@aws-sdk/protocol-http": "^3.40.0",
+    "@aws-sdk/signature-v4": "^3.40.0",
+    "got": "^11.8.2",
+    "p-map": "^4.0.0",
     "source-map-support": "^0.5.16",
     "type-fest": "^2.5.1",
     "xml2js": "^0.4.23"
@@ -59,6 +67,7 @@
     "@typescript-eslint/eslint-plugin": "^5.1.0",
     "@typescript-eslint/parser": "^5.1.0",
     "aws-cdk": "^1.129.0",
+    "aws-sdk-client-mock": "^0.5.6",
     "babel-jest": "^27.3.1",
     "bufferutil": "^4.0.1",
     "eslint": "^7.32.0",
@@ -69,6 +78,7 @@
     "eslint-plugin-unicorn": "^37.0.1",
     "jest": "26.6.0",
     "lodash": "^4.17.21",
+    "nock": "^13.2.0",
     "ts-node": "^9.0.0",
     "typescript": "^4.4.3",
     "utf-8-validate": "^5.0.2",
